+++ airbyte-webapp-e2e-tests/cypress/integration/connection.spec.ts
@@ -17,7 +17,7 @@ describe("Connection main actions", () => {
   });
 
   it("Update connection", () => {
-    cy.intercept("/api/v1/web_backend/connections/updateNew").as("updateConnection");
+    cy.intercept("/api/v1/web_backend/connections/update").as("updateConnection");
 
     createTestConnection("Test update connection source cypress", "Test update connection destination cypress");
 
@@ -27,8 +27,8 @@ describe("Connection main actions", () => {
 
     cy.get("div[data-id='replication-step']").click();
 
-    cy.get("div[data-testid='schedule']").click();
-    cy.get("div[data-testid='Every 5 minutes']").click();
+    cy.get("div[data-testid='scheduleData.basicSchedule']").click();
+    cy.get("div[data-testid='Every hour']").click();
     cy.get("button[type=submit]").first().click();
     cy.wait("@updateConnection");
     cy.get("span[data-id='success-result']").should("exist");