+++ airbyte-workers/src/test/java/io/airbyte/workers/internal/AirbyteMessageTrackerTest.java
@@ -41,7 +41,8 @@ class AirbyteMessageTrackerTest {
 
   @BeforeEach
   void setup() {
-    this.messageTracker = new AirbyteMessageTracker(mStateDeltaTracker, mStateAggregator);
+    final StateMetricsTracker stateMetricsTracker = new StateMetricsTracker(10L * 1024L * 1024L);
+    this.messageTracker = new AirbyteMessageTracker(mStateDeltaTracker, mStateAggregator, stateMetricsTracker);
   }
 
   @Test
@@ -58,7 +59,7 @@ class AirbyteMessageTrackerTest {
 
     assertEquals(3, messageTracker.getTotalRecordsEmitted());
     assertEquals(3L * Jsons.getEstimatedByteSize(r1.getRecord().getData()), messageTracker.getTotalBytesEmitted());
-    assertEquals(2, messageTracker.getTotalStateMessagesEmitted());
+    assertEquals(2, messageTracker.getTotalSourceStateMessagesEmitted());
   }
 
   @Test